+++ code/FinanceManager.Domain/MoneyFlow/Entities/TimeWeightedReturnStatus.cs
@@ -1,0 +1,9 @@
+namespace FinanceManager.Domain.MoneyFlow.Entities;
+
+/// <summary>Explains whether a time-weighted return could be calculated.</summary>
+public enum TimeWeightedReturnStatus
+{
+    Available,
+    InsufficientData,
+    Unavailable,
+}